dealii - add gpu version with kokkos
