+++ examples/deal.II/CMakeLists.txt
@@ -11,13 +11,21 @@ IF(NOT ${deal.II_FOUND})
     )
 ENDIF()
 
-DEAL_II_INITIALIZE_CACHED_VARIABLES()
-PROJECT("bps")
+FILE(GLOB SOURCE_FILES "*.cc")
 
-DEAL_II_INITIALIZE_CACHED_VARIABLES()
+FOREACH ( source_file ${SOURCE_FILES} )
+  GET_FILENAME_COMPONENT(file_name ${source_file} NAME)
+  STRING( REPLACE ".cc" "" exec ${file_name} )
 
-ADD_EXECUTABLE(bps bps.cc)
-DEAL_II_SETUP_TARGET(bps)
+  DEAL_II_INITIALIZE_CACHED_VARIABLES()
+  PROJECT(${exec})
 
-TARGET_INCLUDE_DIRECTORIES(bps PUBLIC ${CEED_DIR}/include)
-TARGET_LINK_LIBRARIES(bps ${CEED_DIR}/lib/libceed.so)
+  DEAL_II_INITIALIZE_CACHED_VARIABLES()
+
+  ADD_EXECUTABLE(${exec} ${source_file})
+  DEAL_II_SETUP_TARGET(${exec})
+
+  TARGET_INCLUDE_DIRECTORIES(${exec} PUBLIC ${CEED_DIR}/include)
+  TARGET_LINK_LIBRARIES(${exec} ${CEED_DIR}/lib/libceed.so)
+
+ENDFOREACH ( source_file ${SOURCE_FILES} )
