+++ examples/deal.II/README.md
@@ -12,10 +12,14 @@ cmake ../ -DDEAL_II_DIR=~/path/to/dealii -DCEED_DIR=~/path/to/libceed
 make
 ```
 
-To run the executable, write:
+To run the executables, write:
 
 ```
-./bps
+./bps_cpu
+```
+
+```
+./bps_kokkos
 ```
 
 Optional command-line arguments are shown by adding the command-line argument "--help".
